[PATCH] 64-bit sector_t - printk changes and sector_t cleanup
From Peter Chubb

printk changes: A sector_t can be either 64 or 32 bits, so cast it to a
printable type that is at least as large as 64-bits on all platforms
(i.e., cast to unsigned long long and use a %llu format)

Transition to 64-bit sector_t: fix isofs_get_blocks by converting the
(possibly 64-bit) arg to a long.

SCSI 64-bit sector_t cleanup: capacity now stored as sector_t; make
sure that the READ_CAPACITY command doesn't sign-extend its returned
value; avoid 64-bit division when printing size in MB.

Still to do:
 - 16-byte SCSI commands
 - Individual scsi drivers.

+++ fs/jbd/commit.c
@@ -355,8 +355,8 @@ sync_datalist_empty:
 			}
 			
 			bh = jh2bh(descriptor);
-			jbd_debug(4, "JBD: got buffer %ld (%p)\n",
-				bh->b_blocknr, bh->b_data);
+			jbd_debug(4, "JBD: got buffer %llu (%p)\n",
+				(unsigned long long)bh->b_blocknr, bh->b_data);
 			header = (journal_header_t *)&bh->b_data[0];
 			header->h_magic     = htonl(JFS_MAGIC_NUMBER);
 			header->h_blocktype = htonl(JFS_DESCRIPTOR_BLOCK);
